% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{add_gnomAD_AF,data.table-method}
\alias{add_gnomAD_AF,data.table-method}
\title{Add AF to MAE counts}
\usage{
\S4method{add_gnomAD_AF}{data.table}(object, ...)
}
\arguments{
\item{object}{a data.frame containing allelic counts.}

\item{list}{... any other parameters such as 'max_af_cutoff', or 'populations' within gnomAD}
}
\value{
a data.frame containing allelic counts and minor allele frequencies
}
\description{
appending the minor allele frequency to allelic counts using gnomAD
}
\examples{
file <- system.file("extdata", "allelic_counts_hg00187.csv", package = "tmae", mustwork = true)
maecounts <- fread(file)
maeres <- deseq4mae(maecounts)

# define the assembly/mafdb you want e.g. hg19, mafdb.gnomad.r2.1.hs37d5, or mafdb.exac.r1.0.hs37d5
\dontrun{
genome_assembly <- 'hg19' 
maeres <- add_gnomad_af(maecounts, genome_assembly = genome_assembly, pop="af")
}
}
