% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{add_gnomAD_AF,GRanges-method}
\alias{add_gnomAD_AF,GRanges-method}
\title{Add AF to GRanges}
\usage{
\S4method{add_gnomAD_AF}{GRanges}(object, ...)
}
\arguments{
\item{object}{a GRanges object}

\item{list}{... any other parameters such as 'max_af_cutoff', or 'populations' within gnomAD}
}
\value{
a data.frame containing GRanges data as well as the minor allele frequencies
}
\description{
appending the minor allele frequency to GRanges using gnomAD
}
