% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{add_gnomAD_AF}
\alias{add_gnomAD_AF}
\title{add gnomAD Frequencies to different datasets}
\usage{
\S4method{add_gnomAD_AF}{data.table}(
  object,
  genome_assembly = c("hg19", "hs37d5", "hg38", "GRCh38"),
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)
}
\description{
add gnomAD Frequencies to different datasets
}
\examples{
file <- system.file("extdata", "allelic_counts_HG00187.csv", package = "tMAE", mustWork = TRUE)
maeCounts <- fread(file)
maeRes <- DESeq4MAE(maeCounts)

# define the assembly/MafDb you want e.g. hg19, MafDb.gnomAD.r2.1.hs37d5, or MafDb.ExAC.r1.0.hs37d5
\dontrun{
genome_assembly <- 'hg19' 
maeRes <- add_gnomAD_AF(maeCounts, genome_assembly = genome_assembly, pop="AF")
}
}
