% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeq2ForMAE.R
\name{allelic_granges_to_dt}
\alias{allelic_granges_to_dt}
\title{Convert GRanges object to a Data Table}
\usage{
allelic_granges_to_dt(data)
}
\arguments{
\item{data}{GRanges object}
}
\value{
data
}
\description{
Converts a GRanges object to a data.table
}
\examples{
allelic_granges_to_dt()
}
