% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeq2ForMAE.R
\name{DESeq4MAE}
\alias{DESeq4MAE}
\title{Run deseq test for MAE}
\usage{
DESeq4MAE(data, minCoverage = 10, disp = 0.05,
  independentFiltering = FALSE)
}
\arguments{
\item{data}{GRanges object.}

\item{disp}{Gene dispersion for the NB test. Default is 0.05.}

\item{independentFiltering}{Parameter that affects the multiple testing. Default is FALSE.}

\item{min_cov}{minimum total allelic count. Default is 10.}
}
\value{
Results.
}
\description{
This function allows you to ...
}
\examples{
file <- system.file("extdata", "demo_MAE_counts.tsv", package = "tMAE", mustWork = TRUE)
maeCounts <- read.table(file)
run_deseq_all_mae(maeCounts)
}
