% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{merge_scores}
\alias{merge_scores}
\title{Add allele frequencies from gnomAD from GRanges}
\usage{
merge_scores(
  data,
  scores,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  max_af_cutoff = 0.001,
  ...
)
}
\arguments{
\item{data}{A data.table object}

\item{scores}{A data.table object of frequency scores generated from gnomAD}

\item{populations}{The population to be annotated.}

\item{max_af_cutoff}{cutoff for a variant to be considered rare. Default is .001.}
}
\value{
A data.table with the original contents plus columns containing allele frequencies from different gnomAD populations.
}
\description{
merge and append information from gnomAD and a data table
}
\author{
Vicente Yepez
}
