% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMA.R
\name{plotMA}
\alias{plotMA}
\title{plotMA}
\usage{
plotMA(data, title = NULL, padjCutoff = 0.05,
  allelicRatioCutoff = 0.8)
}
\arguments{
\item{data}{A data.frame containing the results table from \DESeq4MAE function}

\item{title}{The plot's title}

\item{padjCutoff}{The significance level}

\item{allelicRatioCutoff}{The minimum allelic ratio ALT/(ALT+REF) to be 
considered signficant}
}
\value{
A ggplot object containing the MA plot.
}
\description{
Creates an MA plot, ie, Fold Change (ALT/REF) vs Coverage colored 
   by significance (from p adjusted and allelic ratio)
}
\examples{
plotMA()
}
