% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeq2ForMAE.R
\name{run_deseq_all_mae}
\alias{run_deseq_all_mae}
\title{Run deseq test for MAE}
\usage{
run_deseq_all_mae(data, min_cov = 10, disp = 0.05,
  indepFilter = FALSE, min_fc = log2(3/1), padj_threshold = 0.1)
}
\arguments{
\item{data}{GRanges object.}

\item{min_cov}{Default is 10}

\item{disp}{Default is 0.05}

\item{indepFilter}{Default is False}

\item{min_fc}{Default is log2(3/1)}

\item{padj_threshold}{Default is 0.01}
}
\value{
Results.
}
\description{
This function allows you to ...
}
\examples{
run_deseq_all_mae()
}
