% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeq2ForMAE.R
\name{run_deseq_for_allele_specific_expression}
\alias{run_deseq_for_allele_specific_expression}
\title{Run Deseq for allele Specific Expression}
\usage{
run_deseq_for_allele_specific_expression(data, min_cov = 10,
  disp = 0.05, indepFilter = FALSE)
}
\arguments{
\item{data}{data.table object.}

\item{min_cov}{Default is 10.}

\item{disp}{Default is 0.05}

\item{indepFilter}{Default is False}
}
\value{
A list with the results.
}
\description{
This function allows you to ...
}
\examples{
run_deseq_for_allele_specific_expression()
}
