% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_gnomAD_GR.R
\name{score_gnomAD_GR}
\alias{score_gnomAD_GR}
\title{Add allele frequencies from gnomAD}
\usage{
score_gnomAD_GR(
  gr,
  genome_assembly = c("hg19", "hs37d5", "hg38", "GRCh38"),
  max_af_cutoff = 0.001,
  pops = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)
}
\arguments{
\item{gr}{GenomicRanges object}

\item{genome_assembly}{either 'hg19/hs37d5' or 'hg38/GRCh38' indicating the genome assembly of the variants.
It can also be any full string of a MafDb provided by 
\code{\link[GenomicScores]{availableGScores}}.}

\item{max_af_cutoff}{cutoff for a variant to be considered rare. Default is .001.}

\item{pops}{The population to be annotated.}
}
\value{
A data.table with the scores of each position in the gr object using the mafdb database
}
\description{
Helper Method for add_gnomAD_AF.R generates the gnomAD scoring table given a GenomicRanges object.
}
\examples{
maeCounts <- fread(file)
maeRes <- DESeq4MAE(maeCounts)

# define the assembly/MafDb you want e.g. hg19, MafDb.gnomAD.r2.1.hs37d5, or MafDb.ExAC.r1.0.hs37d5
\dontrun{
genome_assembly <- 'hg19' 
maeRes <- add_gnomAD_AF(maeCounts, genome_assembly = genome_assembly, pop="AF")
}

}
\author{
Vicente Yepez
}
