% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{add_gnomAD_AF,data.table-method}
\alias{add_gnomAD_AF,data.table-method}
\title{Add AF to MAE counts}
\usage{
\S4method{add_gnomAD_AF}{data.table}(object, ...)
}
\arguments{
\item{object}{a data.frame containing allelic counts.}

\item{list}{... any other parameters such as 'max_af_cutoff', or 'populations' within gnomAD}
}
\value{
a data.frame containing allelic counts and minor allele frequencies
}
\description{
Add AF to MAE counts
}
