% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{add_gnomAD_AF,data.table-method}
\alias{add_gnomAD_AF,data.table-method}
\title{Add AF to MAE counts}
\usage{
\S4method{add_gnomAD_AF}{data.table}(
  object,
  genome_assembly = "hg19",
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)
}
\arguments{
\item{object}{data.table containing allelic counts}

\item{genome_assembly}{one of "hg19", "hs37d5", "hg38", "GRCh38" 
It can also be any full string of a MafDb provided by 
\code{\link[GenomicScores]{availableGScores}}.}

\item{max_af_cutoff}{cutoff for a variant to be considered rare. Default 0.001}

\item{populations}{The populations to be annotatated.}

\item{...}{Used for backwards compatibility (gene_assembly -> genome_assembly)}
}
\value{
a data.frame containing GRanges data as well as the minor allele frequencies
}
\description{
appending the minor allele frequency to MAE counts using gnomAD
}
