% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-definitions.R
\name{add_gnomAD_AF}
\alias{add_gnomAD_AF}
\title{add gnomAD AF to object}
\usage{
add_gnomAD_AF(object, ...)
}
\arguments{
\item{object}{data.table or GRanges object}

\item{...}{Further arguments such as populations, max_af_cutoff, or genome_assembly}
}
\value{
A data.frame containing the added gnomAD MAF
}
\description{
appending the minor allele frequency to GRanges using gnomAD
}
