% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{add_gnomAD_AF}
\alias{add_gnomAD_AF}
\title{Add allele frequencies from gnomAD}
\usage{
add_gnomAD_AF(data, gene_assembly = c("hg19", "hg38"),
  max_af_cutoff = 0.001, populations = c("AF", "AF_afr", "AF_amr",
  "AF_eas", "AF_nfe"))
}
\arguments{
\item{data}{A data.frame containing allelic counts.}

\item{gene_assembly}{either 'hg19' or 'hg38' indicating the genome assembly of the variants.}

\item{max_af_cutoff}{cutoff for a variant to be considered rare. Default is .001.}

\item{populations}{}
}
\value{
A data.table with the original contents plus columns containing allele frequencies from different gnomAD populations.
}
\description{
Add allele frequency information from gnomAD.
}
\examples{
# file <- system.file("extdata", "demo_MAE_counts.tsv", package = "tMAE", mustWork = TRUE)
# maeCounts <- read.table(file)
# maeRes <- run_deseq_all_mae(maeCounts)
# maeRes <- add_gnomAD_AF(maeCounts, gene_assembly = 'hg19')
}
\author{
Vicente Yepez
}
