% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gnomAD_AF.R
\name{score_data_GR}
\alias{score_data_GR}
\alias{score_data}
\alias{add_gnomAD_AF<-,GRanges-method}
\title{Add allele frequencies from gnomAD from GRanges}
\usage{
score_data_GR(
  object,
  genome_assembly = c("hg19", "hs37d5", "hg38", "GRCh38"),
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)

score_data(
  object,
  genome_assembly = c("hg19", "hs37d5", "hg38", "GRCh38"),
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)

\S4method{add_gnomAD_AF}{GRanges}(
  object,
  genome_assembly = c("hg19", "hs37d5", "hg38", "GRCh38"),
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
) <- value
}
\arguments{
\item{object}{A data.frame containing allelic counts.}

\item{genome_assembly}{either 'hg19/hs37d5' or 'hg38/GRCh38' indicating the genome assembly of the variants.
It can also be any full string of a MafDb provided by 
\code{\link[GenomicScores]{availableGScores}}.}

\item{max_af_cutoff}{cutoff for a variant to be considered rare. Default is .001.}

\item{populations}{The population to be annotated.}

\item{...}{Used for backwards compatibility (gene_assembly -> genome_assembly)}

\item{gr}{A GRanges object to overlap with gnomAD}
}
\value{
A data.table with the original contents plus columns containing allele frequencies from different gnomAD populations.

A data.table with the original contents plus columns containing allele frequencies from different gnomAD populations.
}
\description{
Add allele frequency information from gnomAD.

Add allele frequency information from gnomAD.
}
\author{
Vicente Yepez

Vicente Yepez
}
