% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeq2ForMAE.R
\name{get_allele_specific_deseq_results}
\alias{get_allele_specific_deseq_results}
\title{Get Allele Specific deseq results}
\usage{
get_allele_specific_deseq_results(dds_res, min_fc = log2(3/1),
  padj_threshold = 0.1)
}
\arguments{
\item{dds_res}{object}

\item{min_fc}{Default is log2(3/1)}

\item{padj_threshold}{Default is 0.1}
}
\value{
data.table object
}
\description{
This function allows you to ...
}
\examples{
get_allele_specific_deseq_results()
}
